+++ gemalaya/qml/HeadingItem.qml
@@ -17,24 +17,31 @@ Text {
   KeyNavigation.tab: nextLinkItem
 
   Layout.margins: 5
-  Layout.maximumWidth: width
   Layout.alignment: Qt.AlignHCenter
+  Layout.fillWidth: true
+  horizontalAlignment: Text.AlignHCenter
 
   signal focusRequested()
 
   TextMetrics {
     id: textmn
-    font.underline: true
+    font.underline: activeFocus || hsize === 'h1'
     font.pointSize: {
       switch(hsize) {
         case 'h1':
-          return Conf.fontPrefs.defaultPointSize * 1.5
+          return activeFocus ?
+                 Conf.fontPrefs.defaultPointSize * 1.6 :
+                 Conf.fontPrefs.defaultPointSize * 1.5
         case 'h2':
-          return Conf.fontPrefs.defaultPointSize * 1.3
+          return activeFocus ?
+                 Conf.fontPrefs.defaultPointSize * 1.4 :
+                 Conf.fontPrefs.defaultPointSize * 1.3
         case 'h3':
-          return Conf.fontPrefs.defaultPointSize * 1.1
+          return activeFocus ?
+                 Conf.fontPrefs.defaultPointSize * 1.2 :
+                 Conf.fontPrefs.defaultPointSize * 1.1
         default:
-          return 16
+          return Conf.fontPrefs.defaultPointSize
       }
     }
     font.bold: true
@@ -44,11 +51,11 @@ Text {
   color: {
     switch(hsize) {
       case 'h1':
-        return Conf.heading.h1.color
+        return activeFocus ? Conf.heading.h1.colorFocused : Conf.heading.h1.color
       case 'h2':
-        return Conf.heading.h2.color
+        return activeFocus ? Conf.heading.h2.colorFocused : Conf.heading.h2.color
       case 'h3':
-        return Conf.heading.h3.color
+        return activeFocus ? Conf.heading.h3.colorFocused : Conf.heading.h3.color
       default:
         return 'white'
     }
@@ -58,10 +65,4 @@ Text {
   font: textmn.font
   antialiasing: true
   wrapMode: Text.WrapAnywhere
-
-  onFocusChanged: {
-    if (focus && nextLinkItem) {
-      nextLinkItem.forceActiveFocus()
-    }
-  }
 }